Revert "Moved back xcode post copy script to use cp since rsync not working"
This reverts commit 3b9c1e6584a6b2d3ad447b790e93323c57156f8b.

+++ gameplay/src/Model.cpp
@@ -310,7 +310,7 @@ static bool drawWireframe(MeshPart* part)
     {
     case Mesh::TRIANGLES:
         {
-            for (size_t i = 0; i < indexCount; i += 3)
+            for (unsigned int i = 0; i < indexCount; i += 3)
             {
                 GL_ASSERT( glDrawElements(GL_LINE_LOOP, 3, part->getIndexFormat(), ((const GLvoid*)(i*indexSize))) );
             }
@@ -319,7 +319,7 @@ static bool drawWireframe(MeshPart* part)
 
     case Mesh::TRIANGLE_STRIP:
         {
-            for (size_t i = 2; i < indexCount; ++i)
+            for (unsigned int i = 2; i < indexCount; ++i)
             {
                 GL_ASSERT( glDrawElements(GL_LINE_LOOP, 3, part->getIndexFormat(), ((const GLvoid*)((i-2)*indexSize))) );
             }
